+++ packages/spotlight/CHANGELOG.md
@@ -14,6 +14,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight` to re-focus the last-focused element when window regains focus
+
 ## [1.6.1] - 2017-08-07
 
 No significant changes.